feat!: refactoring
Moved functions into `docgen::` namespace.

Replaced `DOCGEN_DOC_MARK` with `DOCGEN_MARK` and
changed default to `##`, to fix issues with external formatters
removing trailing whitespaces. This allows for inclusion of empty
lines in documentation, which was an issue with defninition lists.

Moved most code into `main` function, used a different check for
execution/sourcing of the file.

Used `cat` as pager for `man` to get rid of hang-ups waiting for
user input.